足迹

+++ PeachTravel/libs/SDSegmentedControl/SDSegmentedControl.m
@@ -448,7 +448,7 @@ const CGFloat kSDSegmentedControlScrollOffset = 20;
 - (NSInteger)firstEnabledSegmentIndexNearIndex:(NSUInteger)index
 {
     // Select the first enabled segment
-    for (int i = index; i < self._items.count; i++)
+    for (NSInteger i = index; i < self._items.count; i++)
     {
         if (((SDSegmentView *)self._items[i]).enabled)
         {
@@ -456,7 +456,7 @@ const CGFloat kSDSegmentedControlScrollOffset = 20;
         }
     }
 
-    for (int i = index; i >= 0; i--)
+    for (NSInteger i = index; i >= 0; i--)
     {
         if (((SDSegmentView *)self._items[i]).enabled)
         {
